add toggle command
The Liebert MPX PDU does not have native support for a toggle command,
but we can emulate it by caching the port state. This is useful to offer
for use-cases like toggle buttons. By offering this directly in the MQTT
API they do not need to track the port state.

Note, that this command is racy. The PDU needs some time to change its
state. If two toggle commands arrive quickly after each other the
internal state has not yet changed and the port will change its state
only once. It would also be possible to update the internal state
immediately, but then toggle is faster then the PDU, which is also
weird.